Added Direct App loading for Samsung TV
Discovered how to load Samsung Smart TV Apps directly over HTTP POST and you can also do GET to know what App is currently loaded. I used to do this with a complex sequence of IR commands which works most of the time but whenever Samsung push an update those Apps are now in a new location and it breaks, it also is alot quicker to call the app rather then a long winded IR script.